fix: Properly output conflict resolutions for shorthands. (#238)
- ConflictResolver takes shorthands into consideration.
- PropertyParser has special cases for border and font-family because of bugs in source package.
- fix: Export syntax now allows for immediate re-export from file path.
- Added more debug statements.
- Fix Broccoli bug to only remove files from tmp directory output.

+++ packages/@css-blocks/broccoli/src/Analyze.ts
@@ -111,7 +111,7 @@ export class CSSBlocksAnalyze extends BroccoliPlugin {
         let filename = filesystemPath || options.importer.debugIdentifier(block.identifier, options);
 
         // If this Block has a representation on disk, remove it from our output tree.
-        if (filesystemPath) {
+        if (filesystemPath && !!~filesystemPath.indexOf(output)) {
           debug(`Removing block file ${filesystemPath} from output.`);
           if (fs.existsSync(filesystemPath)) { fs.removeSync(filesystemPath); }
         }
